Version 3.0.0
-Fixed bug that prevented correct detection of Mc Version
